Settings tab
A new tab has been added in Booking->Settings for the addon. This tab
has 2 checkboxes that will allow the admin to save the settings and
enable/disable the addon.

The settings added are as below:
1. Enable a user to send booking details to a Friend: This will allow
the Book Another Slot and Send a Friend button to be displayed on the
Thank You page and order emails.

2. Mark the admin in cc in emails sent to friends: This will
enable/disable the marking of admin in cc in the emails that are sent to
the friends. When the setting is disabled, only the user (who has sent
the email) will be marked.