Fix #9
The issue was occurring when we use the 'shop' page url as the base. I
have now changed it to use the 'cart' page url. It works fine with both
the default permalink as well as a custom permalink ( have checked it
only 1 custom link). Please make sure it is tested with the others as
well.