Strict Standards Error
When calling the function get_item_meta() using the scope resolution
operator as WC_Abstract_Order::get_item_meta( $item_id ), a strict
standards error is displayed on the Product page as the method is not
defined as static. Since, the class is an abstract class, we cannot
create an object of the class either. To fix this issue, I have now used
the the wc_get_order_item_meta function which can be used to get the
value of individual item meta keys.