Creating virtual page for "Tell a friend" page so it loads correctly for all themes.
Creating virtual page for "Tell a friend" page so it loads correctly for
all themes.
The "template_include" has been removed. The reason is that it is used
to load a particular template. We were actually using it to load a page.
Creating a virtual page temporarily is the best solution I found so the
page loads like any other page which is created from the "Pages"
section.