Use the slug parameter and create a new page for the addon
These changes will now ensure that a new page is generated for the 'Tell
a friend' page. Earlier we were using the 'shop' or 'cart' pages to
introduce our page and passing parameters in the $_GET variable to
identify the page. We now use the slug parameter to achieve the same.
However, this needs to be tested with different permalinks setting to
ensure it works with all.