Fix #25
The below fixes ensure that the qty and add to cart buttons are
displayed on the variable product page. The issue arises because
Woocommerce adds this buttons on the page only after a variation is
selected in the drop down OR the product has a default variation
selected. This has been fixed by calling the function
"woocommerce_single_variation_add_to_cart_button"