Fix #9
A setting has been added in the Booking->Settings->Addon Settings tab.
Admin can setup the url that he would like to use for the 'Tell a
Friend' page. the default value is send-booking-to-friend

The addon works with all Permalink Settings except Default and Numeric
and should be tested accordingly.