Issue fixes
The Tell a Friend page was not loading when the url was changed in the
settings. This has been fixed.

The default css for the buttons has been modified to make the button
text appear in the middle of the button

The hidden field which lists the products selected on the Tell a friend
page has been modified to list only those products that are available.