Fix #61, Fix #62, Fix #63, Fix #64
The plugin name was modified on wordpress.org. As a result the plugin
folder name was also automatically updated. Modified the folder name in
the file to ensure it is detected.

The labels used when sending the emails are incorrect. Modified them to
use the labels on the Cart and Checkout page, instead of the Product
page.